Conflicts:
	packages/zoltan2/sphynx/src/Zoltan2_Sphynx.hpp
	packages/zoltan2/sphynx/src/Zoltan2_SphynxProblem.hpp

+++ packages/stk/stk_mesh/stk_mesh/baseImpl/elementGraph/ElemGraphShellConnections.cpp
@@ -98,6 +98,7 @@ bool SideConnections::does_side_have_both_connection_to_shell_and_to_nonshell(in
 
 void fill_non_shell_graph_edges_to_delete(GraphInfo &graphInfo, const stk::mesh::impl::ElementSidePair &elementSidePair, std::vector<GraphEdge>& edgesToDelete)
 {
+    std::vector<GraphEdge> pllEdgesToDelete;
     for(const stk::mesh::GraphEdge& graphEdge : graphInfo.graph.get_edges_for_element_side(elementSidePair.first, elementSidePair.second))
     {
         bool thisIsShell = is_this_element_shell(graphEdge, graphInfo.elementTopologies);
@@ -105,11 +106,13 @@ void fill_non_shell_graph_edges_to_delete(GraphInfo &graphInfo, const stk::mesh:
 
         if(!thisIsShell && !thatIsShell)
         {
-            if(!impl::is_local_element(graphEdge.elem2()))
-                graphInfo.parGraphInfo.erase_parallel_info_for_graph_edge(graphEdge);
+            if(!impl::is_local_element(graphEdge.elem2())) {
+                pllEdgesToDelete.push_back(graphEdge);
+            }
             edgesToDelete.push_back(graphEdge);
         }
     }
+    graphInfo.parGraphInfo.erase_edges(pllEdgesToDelete);
 }
 
 void remove_graph_edges_blocked_by_shell(GraphInfo &graphInfo)
@@ -120,8 +123,14 @@ void remove_graph_edges_blocked_by_shell(GraphInfo &graphInfo)
         SideConnections sideConnectionsForElement(graphInfo.elementTopologies[localId].num_sides());
         for(int side : sideConnectionsForElement.get_sides_connected_to_shell_and_nonshell(graphInfo, localId))
             fill_non_shell_graph_edges_to_delete(graphInfo, stk::mesh::impl::ElementSidePair(localId, side), edgesToDelete);
+
+        if (edgesToDelete.size() > 0)
+        {
+            std::sort(edgesToDelete.begin(), edgesToDelete.end(), GraphEdgeLessByElem1());
+            graphInfo.graph.delete_sorted_edges(edgesToDelete);
+            edgesToDelete.clear();
+        }
     }
-    graphInfo.graph.delete_sorted_edges(edgesToDelete);
 }
 
 }
